+++ php-book-app/app/src/Controller/QuestionsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace  App\Controller;
+namespace App\Controller;
 
 /**
  * Questions Controller
@@ -23,7 +23,7 @@ class QuestionsController extends AppController
      */
     public function index()
     {
-        $questions = $this->paginate($this->Questions->findQuestionsWithAnsweredCount(), [
+        $questions = $this->paginate($this->Questions->findQuestionsWithAnsweredCount()->contain(['Users']), [
             'order' => ['Questions.id' => 'DESC']
         ]);
 
@@ -33,15 +33,16 @@ class QuestionsController extends AppController
     /**
      * 質問投稿画面/質問投稿処理
      *
-     * @return \Cake\Http\Response|null 質問投稿後に質問一覧画面に遷移する
+     * @return \Cake\Http\Response|null 質問投稿後に質問一覧画面へ遷移する
      */
     public function add()
     {
         $question = $this->Questions->newEntity();
 
         if ($this->request->is('post')) {
-            $question = $this->Questions->pathEntity($question, $this->request->getData());
-            $question->user_id = 1; //todoユーザー管理機能実装時に修正する
+            $question = $this->Questions->patchEntity($question, $this->request->getData());
+
+            $question->user_id = $this->Auth->user('id');
 
             if ($this->Questions->save($question)) {
                 $this->Flash->success('質問を投稿しました');
@@ -58,16 +59,17 @@ class QuestionsController extends AppController
      * 質問詳細画面
      *
      * @param int $id 質問ID
-     * @return \Cake\Http\Response|null 質問投稿後に質問一覧画面に遷移する
+     * @return void
      */
     public function view(int $id)
     {
-        $question = $this->Questions->get($id);
+        $question = $this->Questions->get($id, ['contain' => ['Users']]);
 
         $answers = $this
             ->Answers
             ->find()
             ->where(['Answers.question_id' => $id])
+            ->contain(['Users'])
             ->orderAsc('Answers.id')
             ->all();
 
@@ -80,14 +82,17 @@ class QuestionsController extends AppController
      * 質問削除処理
      *
      * @param int $id 質問ID
-     * @return \Cake\Http\Response|null 質問削除後に質問一覧画面に遷移する
+     * @return \Cake\Http\Response|null 質問削除後に質問一覧画面へ遷移する
      */
     public function delete(int $id)
     {
         $this->request->allowMethod(['post']);
 
         $question = $this->Questions->get($id);
-        //todo 質問を削除できるのは質問投稿者のみとする
+        if ($question->user_id !== $this->Auth->user('id')) {
+            $this->Flash->error('他のユーザーの質問を削除することは出来ません');
+            return $this->redirect(['action' => 'index']);
+        }
 
         if ($this->Questions->delete($question)) {
             $this->Flash->success('質問を削除しました');
